Solucion Estructuras_Control_Flujo

+++ Estructuras_Control_Flujo/Estructuras_Control_Flujo/2_06_AplicaciondeDescuentoCASE.vb
@@ -1,7 +1,7 @@
 ﻿Module AplicaciondeDescuentoCASE
     Dim Cantidad As Integer
-    Dim PrecioUnitario As Double
-    Dim Descuento As Double = 0
+    Dim PrecioUnitario As Single
+    Dim Descuento As Single = 0
     Sub Main()
         Console.WriteLine("Calculo de Descuento")
         Console.Write("Ingrese la cantidad: ")
@@ -18,8 +18,8 @@
                 Descuento = 0.2
         End Select
 
-        Dim SubTotal As Double = Cantidad * PrecioUnitario
-        Dim MontoDescontado As Double = SubTotal * Descuento
+        Dim SubTotal As Single = Cantidad * PrecioUnitario
+        Dim MontoDescontado As Single = SubTotal * Descuento
         Console.WriteLine("SubTotal: $" & SubTotal)
         Console.WriteLine("Porcentaje de Descuento: " & Descuento * 100 & "%")
         Console.WriteLine("Monto Descontado: $" & MontoDescontado)
